// src/shared/api/generated/types/NewsPaginationMetaDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type NewsPaginationMetaDto = {
  /**
   * @type number
   */
  page: number;
  /**
   * @type number
   */
  limit: number;
  /**
   * @type number
   */
  total: number;
  /**
   * @type number
   */
  totalPages: number;
};